requirements.txt
asgiref==3.8.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
Django==5.1.2
djangorestframework==3.15.2
fonttools==4.54.1
idna==3.10
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
reportlab==4.2.5
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
